Fixed "DD image already exists. Delete it first" error
When trying to image as a DD file, a check for the existance of the file
was made, but it was made after first creating it! So it was always true
and therefore wouldn't let you image as DD even if the file did not
actually exist on disk yet. Fixed for now by removing the overwrite
check.